+++ packages/react-reconciler/src/Scheduler.js
@@ -1,0 +1,46 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+// This module only exists as an ESM wrapper around the external CommonJS
+// Scheduler dependency. Notice that we're intentionally not using named imports
+// because Rollup would use dynamic dispatch for CommonJS interop named imports.
+// When we switch to ESM, we can delete this module.
+import * as Scheduler from 'scheduler';
+import {__interactionsRef} from 'scheduler/tracing';
+import {enableSchedulerTracing} from 'shared/ReactFeatureFlags';
+import invariant from 'shared/invariant';
+
+export const scheduleCallback = Scheduler.unstable_scheduleCallback;
+export const cancelCallback = Scheduler.unstable_cancelCallback;
+export const shouldYield = Scheduler.unstable_shouldYield;
+export const requestPaint = Scheduler.unstable_requestPaint;
+export const now = Scheduler.unstable_now;
+export const getCurrentPriorityLevel =
+  Scheduler.unstable_getCurrentPriorityLevel;
+export const ImmediatePriority = Scheduler.unstable_ImmediatePriority;
+export const UserBlockingPriority = Scheduler.unstable_UserBlockingPriority;
+export const NormalPriority = Scheduler.unstable_NormalPriority;
+export const LowPriority = Scheduler.unstable_LowPriority;
+export const IdlePriority = Scheduler.unstable_IdlePriority;
+
+if (enableSchedulerTracing) {
+  // Provide explicit error message when production+profiling bundle of e.g.
+  // react-dom is used with production (non-profiling) bundle of
+  // scheduler/tracing
+  invariant(
+    __interactionsRef != null && __interactionsRef.current != null,
+    'It is not supported to run the profiling version of a renderer (for ' +
+      'example, `react-dom/profiling`) without also replacing the ' +
+      '`scheduler/tracing` module with `scheduler/tracing-profiling`. Your ' +
+      'bundler might have a setting for aliasing both modules. Learn more at ' +
+      'https://reactjs.org/link/profiling',
+  );
+}
+
+export type SchedulerCallback = (isSync: boolean) => SchedulerCallback | null;
